+++ wetterdienst/metadata/kind.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2018-2021, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from enum import Enum
+
+
+class Kind(Enum):
+    OBSERVATION = "observation"
+    FORECAST = "forecast"
+    RADAR = "radar"