+++ wetterdienst/util/geo.py
@@ -76,6 +76,6 @@ def convert_dm_to_dd(dms: float) -> float:
     """ Convert degree minutes to decimal degree """
     degrees, minutes = divmod(dms, 1)
 
-    decimals = round(minutes * (60 / 100), 2)
+    decimals = round(minutes / 60 * 100, 2)
 
     return degrees + decimals